Clarify format used in insertImage
This change makes the format used by insertImage() explicit, allowing code that requires this information to depend on it symbolically rather than by examining the source. A case where this is useful is when specifying the MIME type in Intent.setType() when passing the result of insertImage as an Intent extra and attempting to invoke an activity that handles most "image/*" MIME types but that does not have a registered intent filter for "vnd.android.cursor.dir/image". In such a scenario, code might have hard-coded "image/jpeg" as a string; after this change, it becomes possible to use the STORED_CONTENT_TYPE variable, which would make such code less fragile.